+++ src/main/java/calculator/domain/StringCalculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 public class StringCalculator {
