+++ src/main/java/calculator/domain/operation/AddOperation.java
@@ -1,4 +1,4 @@
-package calculator.operation;
+package calculator.domain.operation;
 
 public class AddOperation implements Operation {
