+++ src/main/java/calculator/domain/operation/MinusOperation.java
@@ -1,4 +1,4 @@
-package calculator.operation;
+package calculator.domain.operation;
 
 public class MinusOperation implements Operation {
