+++ src/main/java/calculator/domain/operation/MultiplyOperation.java
@@ -1,4 +1,4 @@
-package calculator.operation;
+package calculator.domain.operation;
 
 public class MultiplyOperation implements Operation {
