+++ src/main/java/calculator/domain/operation/Operation.java
@@ -1,4 +1,4 @@
-package calculator.operation;
+package calculator.domain.operation;
 
 public interface Operation {
