+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.domain.StringCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
